requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
backports-abc==0.5
bleach==3.1.0
blessings==1.7
certifi==2019.3.9
chardet==3.0.4
cloudpickle==1.1.1
colorama==0.4.1
cycler==0.10.0
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
gpustat==0.5.0
grpcio==1.21.1
h5py==2.9.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==5.7.8
numpy==1.16.3
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-contrib-python==4.1.0.25
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.11.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
putil==0.9.12
py==1.8.0
PyContracts==1.8.12
Pygments==2.4.1
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.5.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.0
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
tensorboardX==1.7
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
torch==0.4.1
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.2.1
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.25.3
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.4.2
youtube-dl==2019.5.20
